tests/requirements-test.txt
# Test framework
pytest>=8.0
pytest-asyncio>=0.23
pytest-html>=4.0
pytest-timeout>=2.2

# HTTP client for API tests
httpx>=0.27

# Browser automation for E2E tests
playwright>=1.40

# AI evaluation (reuses project's Gemini SDK)
google-genai>=1.0

# Environment variable loading
python-dotenv>=1.0
